Add more methods in POM
