+++ requirements.txt
@@ -17,7 +17,6 @@ packaging==21.0
 pipenv==2021.5.29
 pluggy==1.0.0
 py==1.10.0
-pyparsing==3.0.3
 pytest==6.2.5
 pytest-ordering==0.6
 requests==2.26.0
@@ -28,3 +27,4 @@ virtualenv==20.9.0
 virtualenv-clone==0.5.7
 playwright==1.16.1
 pytest-playwright==0.2.2
+selenium
